+++ CHANGELOG.md
@@ -20,7 +20,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - improved compatibility template for Privacy Enhanced box types (thanks offhub) [#2899](https://github.com/sandboxie-plus/Sandboxie/pull/2899)
+- improved support page in settings and reminder [#2896](https://github.com/sandboxie-plus/Sandboxie/issues/2896)
 - improved signature error message [#2931](https://github.com/sandboxie-plus/Sandboxie/issues/2931)
+- changed Don't show "No Inet" when exceptions exist [#2919](https://github.com/sandboxie-plus/Sandboxie/issues/2919)
 
 ### Fixed
 - fixed Qt6 issues in ARM64 build
@@ -35,7 +37,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - fixed System tray icon hourglass overlay gets stuck when operation is stopped [#2869](https://github.com/sandboxie-plus/Sandboxie/issues/2869)
 - fixed File Panel doesn't allow to adjust columns size in a permanent way [#2930](https://github.com/sandboxie-plus/Sandboxie/issues/2930)
 - fixed Renaming a box with sandboxed run entries can break those entries. [#2921](https://github.com/sandboxie-plus/Sandboxie/issues/2921)
-
+- fixed WFP not enabled after Setup Wizard [#2915](https://github.com/sandboxie-plus/Sandboxie/issues/2915)